refactor relayer utils into bindings and reuse in workers
